refactor: remove redundant facade exports
