+++ fallow-baselines/production-unused-exports.json
@@ -7,16 +7,6 @@
     "src/cloud-webdriver/aws-device-farm.ts:createAwsDeviceFarmWebDriverRuntime",
     "src/cloud-webdriver/browserstack.ts:getBrowserStackWebDriverCapabilities",
     "src/cloud-webdriver/browserstack.ts:createBrowserStackWebDriverRuntime",
-    "src/commands/capture/index.ts:alertCliReader",
-    "src/commands/capture/index.ts:alertDaemonWriter",
-    "src/commands/capture/index.ts:diffCliReader",
-    "src/commands/capture/index.ts:screenshotCliReader",
-    "src/commands/capture/index.ts:screenshotDaemonWriter",
-    "src/commands/capture/index.ts:settingsCliReader",
-    "src/commands/capture/index.ts:settingsDaemonWriter",
-    "src/commands/capture/index.ts:snapshotCliReader",
-    "src/commands/capture/index.ts:waitCliReader",
-    "src/commands/capture/index.ts:waitDaemonWriter",
     "src/commands/command-metadata.ts:listCommandMetadataNames",
     "src/commands/command-surface.ts:listExecutableCommandNames",
     "src/commands/index.ts:ref",
@@ -30,23 +20,7 @@
     "src/core/command-descriptor/registry.ts:listCommandResponseDataTransforms",
     "src/core/dispatch.ts:listRegisteredDispatchCommandNames",
     "src/core/platform-plugin/plugin.ts:registeredPlatforms",
-    "src/daemon/app-log.ts:APP_LOG_PID_FILENAME",
-    "src/daemon/app-log.ts:cleanupStaleAppLogProcesses",
-    "src/daemon/app-log.ts:assertAndroidPackageArgSafe",
-    "src/daemon/app-log.ts:buildAppleLogPredicate",
-    "src/daemon/app-log.ts:buildIosDeviceConsoleLaunchArgs",
-    "src/daemon/app-log.ts:buildIosSimulatorLogStreamArgs",
-    "src/daemon/client/daemon-client.ts:computeDaemonCodeSignature",
-    "src/daemon/client/daemon-client.ts:downloadRemoteArtifact",
-    "src/daemon/client/daemon-client.ts:cleanupFailedDaemonStartupMetadata",
-    "src/daemon/client/daemon-client.ts:resolveDaemonStartupHint",
-    "src/daemon/client/daemon-client.ts:canConnectSocket",
-    "src/daemon/client/daemon-client.ts:resolveDaemonRequestTimeoutMs",
-    "src/daemon/client/daemon-client.ts:shouldResetDaemonAfterRequestTimeout",
-    "src/daemon/handlers/find.ts:parseFindArgs",
-    "src/daemon/handlers/snapshot-capture.ts:buildSnapshotVisibility",
     "src/daemon/lease-context.ts:buildLeaseDiagnosticsContext",
-    "src/daemon/runtime-hints.ts:resolveRuntimeTransportHints",
     "src/kernel/device.ts:isPlatform",
     "src/platforms/android/multitouch-helper.ts:resetAndroidMultiTouchHelperInstallCache",
     "src/platforms/android/snapshot-helper-install.ts:resetAndroidSnapshotHelperInstallCache",
@@ -56,7 +30,6 @@
     "src/platforms/install-source.ts:ARCHIVE_EXTENSIONS",
     "src/platforms/linux/linux-env.ts:resetInputToolCache",
     "src/provider-device-runtime.ts:setActiveProviderDeviceRuntimes",
-    "src/remote/remote-config.ts:resolveRemoteConfigPath",
     "src/utils/ttl-memo.ts:resetAllProcessMemosForTests"
   ],
   "unused_types": [],
