+++ src/commands/capture/index.ts
@@ -1,14 +1,10 @@
 import { defineCommandFamilyFromFacets } from '../family/types.ts';
-import { alertCommandFacet, alertCliReader, alertDaemonWriter } from './alert.ts';
-import { diffCommandFacet, diffCliReader } from './diff.ts';
-import {
-  screenshotCommandFacet,
-  screenshotCliReader,
-  screenshotDaemonWriter,
-} from './screenshot.ts';
-import { settingsCliReader, settingsCommandFacet, settingsDaemonWriter } from './settings.ts';
-import { snapshotCommandFacet, snapshotCliReader } from './snapshot.ts';
-import { waitCommandFacet, waitCliReader, waitDaemonWriter } from './wait.ts';
+import { alertCommandFacet } from './alert.ts';
+import { diffCommandFacet } from './diff.ts';
+import { screenshotCommandFacet } from './screenshot.ts';
+import { settingsCommandFacet } from './settings.ts';
+import { snapshotCommandFacet } from './snapshot.ts';
+import { waitCommandFacet } from './wait.ts';
 
 const captureCommandFacets = [
   snapshotCommandFacet,
@@ -23,16 +19,3 @@ export const captureCommandFamily = defineCommandFamilyFromFacets({
   name: 'capture',
   commands: captureCommandFacets,
 });
-
-export {
-  alertCliReader,
-  alertDaemonWriter,
-  diffCliReader,
-  screenshotCliReader,
-  screenshotDaemonWriter,
-  settingsCliReader,
-  settingsDaemonWriter,
-  snapshotCliReader,
-  waitCliReader,
-  waitDaemonWriter,
-};
